fix: move home to app
